+++ pages/index.js
@@ -96,7 +96,7 @@ function Home({ dictionaries }) {
 
           <style jsx>{`
             main {
-              padding: 5rem 0;
+              padding: 1rem 0 5rem 0;
               flex: 1;
               display: flex;
               flex-direction: column;